spendshield/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.18.0
boto3>=1.26.0
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.21.0
slowapi>=0.1.4
pydantic-settings
